feat: support update embedded-product (v5) Move the code to migrate.go
